# api/bundled/openapi.yaml
openapi: 3.0.3
info:
  title: Moneo Backend API (as-is)
  version: 0.1.0
  license:
    name: Proprietary
    url: https://example.com/license
  description: |
    Contract documents current behavior of auth, catalogs and transactions endpoints as implemented.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: refreshAuth
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
...
components:
  schemas:
...
    AuthResponse:
      type: object
      required:
        - access_token
        - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
    AuthErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
  responses:
    AuthError:
      description: Auth flow error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT